fix bugs & change account changing button location
