+++ src/pages/my-account/my-account.html
@@ -7,6 +7,11 @@
 <ion-header>
 
   <ion-navbar [hideBackButton]="true">
+    <ion-buttons left>
+      <button ion-button ion-long-press (onPressing)="switchUserEditor()" (click)="switchUser()">
+        <ion-icon name="repeat"></ion-icon>
+      </button>
+    </ion-buttons>
     <ion-title>My Account</ion-title>
     <ion-buttons end>
     	<button ion-button (click)="goToItemList()">
@@ -23,4 +28,28 @@
 		<ion-tab [root]="myItemsRoot" tabTitle="My Items"></ion-tab>
 		<ion-tab [root]="activityRoot" tabTitle="Activity" tabBadgeStyle="danger" [tabBadge]="getActivityNumber()"></ion-tab>
 	</ion-tabs>
+
+  <div class="user-info-editor" *ngIf="isUserInfoEditorVisible">
+    <ion-list>
+      <ion-list-header>
+        User #1
+      </ion-list-header>
+      <ion-item no-lines>
+        <ion-input type="text" placeholder="User #1 Name" [(ngModel)]="users.userA.user_name"></ion-input>
+      </ion-item>
+      <ion-item no-lines>
+        <ion-input type="text" placeholder="User #1 Email" [(ngModel)]="users.userA.user_email"></ion-input>
+      </ion-item>
+      <ion-list-header>
+        User #2
+      </ion-list-header>
+      <ion-item no-lines>
+        <ion-input type="text" placeholder="User #2 Name" [(ngModel)]="users.userB.user_name"></ion-input>
+      </ion-item>
+      <ion-item no-lines>
+        <ion-input type="text" placeholder="User #2 Email" [(ngModel)]="users.userB.user_email"></ion-input>
+      </ion-item>
+      <button ion-button block (click)="applyUserInfo()">Apply</button>
+    </ion-list>
+  </div>
 </ion-content>
