+++ src/pages/my-account/my-account.scss
@@ -12,3 +12,16 @@ page-my-account {
 		}
 	}
 }
+
+.user-info-editor{
+  float: left;
+  position: absolute;
+  z-index: 999;
+  background: white;
+  width: 100%;
+  padding: 10px 16px 10px 16px;
+  box-shadow: 0 2px 6px rgba(0, 0, 0, 0.2);
+  ion-list{
+    margin: 0;
+  }
+}
